Fixed the script to work with the new SE topbar
When SE rolled out the new topbar netwerk-wide on 2022-03-30, they changed a lot of the element structure. One of these was the user card which contains the current user reputation. Since the userscript needs the user's reputation at startup, it meant it completely failed to initialise and run.

Changed the selector for the user information to the now correct one for the new structure.